moved missing value function to connectors.hdf5
