+++ tests/test_timeseries.py
@@ -393,32 +393,3 @@ def test_light_attach():
     full = generate_timeseries()
     light = generate_light_timeseries()
     compare_timeseries(light, full, verbose=True)
-
-def test_negative_zero_matrix():
-    """
-    timeseries.negative_zero_matrix()
-    """
-    num_cols = 40
-    num_rows = 800
-    num_missing = num_cols * num_rows / 4
-
-    random.seed(0)
-    dataset = numpy.random.random(size=(num_rows, num_cols)) + 0.1
-    inverse = numpy.full((num_rows, num_cols), False)
-
-    remove_list = set([])
-    for _ in range(num_missing):
-        irow = numpy.random.randint(0, num_rows)
-        icol = numpy.random.randint(0, num_cols)
-        remove_list.add((irow, icol))
-
-    for irow, icol in remove_list:
-        dataset[irow, icol] = -0.0
-        inverse[irow, icol] = True
-
-    missing_matrix = tokio.timeseries.negative_zero_matrix(dataset)
-
-    print "Added %d missing data; missing_matrix contains %d" % (len(remove_list),
-                                                                 missing_matrix.sum())
-    assert len(remove_list) == missing_matrix.sum()
-    assert ((missing_matrix == 0.0) | inverse).all()
